Abstract out tokens from being tied tightly to access
