+++ src/Token/AccessToken.php
@@ -22,7 +22,7 @@ use RuntimeException;
  *
  * @link http://tools.ietf.org/html/rfc6749#section-1.4 Access Token (RFC 6749, §1.4)
  */
-class AccessToken implements AccessTokenInterface, ResourceOwnerAccessTokenInterface
+class AccessToken extends AbstractToken implements AccessTokenInterface, ResourceOwnerAccessTokenInterface
 {
     /**
      * @var string
@@ -44,54 +44,12 @@ class AccessToken implements AccessTokenInterface, ResourceOwnerAccessTokenInter
      */
     protected $resourceOwnerId;
 
-    /**
-     * @var array
-     */
-    protected $values = [];
-
-    /**
-     * @var int
-     */
-    private static $timeNow;
-
-    /**
-     * Set the time now. This should only be used for testing purposes.
-     *
-     * @param int $timeNow the time in seconds since epoch
-     * @return void
-     */
-    public static function setTimeNow($timeNow)
-    {
-        self::$timeNow = $timeNow;
-    }
-
-    /**
-     * Reset the time now if it was set for test purposes.
-     *
-     * @return void
-     */
-    public static function resetTimeNow()
-    {
-        self::$timeNow = null;
-    }
-
-    /**
-     * @return int
-     */
-    public function getTimeNow()
-    {
-        return self::$timeNow ? self::$timeNow : time();
-    }
-
-    /**
-     * Constructs an access token.
-     *
-     * @param array $options An array of options returned by the service provider
-     *     in the access token request. The `access_token` option is required.
-     * @throws InvalidArgumentException if `access_token` is not provided in `$options`.
-     */
+  /**
+   * @inheritDoc
+   */
     public function __construct(array $options = [])
     {
+        parent::__construct($options);
         if (empty($options['access_token'])) {
             throw new InvalidArgumentException('Required option not passed: "access_token"');
         }
@@ -104,27 +62,6 @@ class AccessToken implements AccessTokenInterface, ResourceOwnerAccessTokenInter
 
         if (!empty($options['refresh_token'])) {
             $this->refreshToken = $options['refresh_token'];
-        }
-
-        // We need to know when the token expires. Show preference to
-        // 'expires_in' since it is defined in RFC6749 Section 5.1.
-        // Defer to 'expires' if it is provided instead.
-        if (isset($options['expires_in'])) {
-            if (!is_numeric($options['expires_in'])) {
-                throw new \InvalidArgumentException('expires_in value must be an integer');
-            }
-
-            $this->expires = $options['expires_in'] != 0 ? $this->getTimeNow() + $options['expires_in'] : 0;
-        } elseif (!empty($options['expires'])) {
-            // Some providers supply the seconds until expiration rather than
-            // the exact timestamp. Take a best guess at which we received.
-            $expires = $options['expires'];
-
-            if (!$this->isExpirationTimestamp($expires)) {
-                $expires += $this->getTimeNow();
-            }
-
-            $this->expires = $expires;
         }
 
         // Capture any additional values that might exist in the token but are
